Add BaseActivity
Add BaseActivity to hold the navigation bar.
