+++ MyApplication/app/src/main/java/androidstudio/edbud/com/myapplication/CoursePage.java
@@ -21,7 +21,7 @@ import android.widget.ListView;
 
 import java.util.ArrayList;
 
-public class CoursePage extends AppCompatActivity implements View.OnClickListener, NavigationView.OnNavigationItemSelectedListener {
+public class CoursePage extends BaseActivity implements View.OnClickListener, NavigationView.OnNavigationItemSelectedListener {
     public static ArrayList courses = new ArrayList();
     public static ArrayList units = new ArrayList();
     public static ArrayList<courses> myCourse = new ArrayList<courses>();
@@ -36,17 +36,7 @@ public class CoursePage extends AppCompatActivity implements View.OnClickListene
         super.onCreate(savedInstanceState);
         context = this;
         setContentView(R.layout.activity_course_page);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout1);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.setDrawerListener(toggle);
-        toggle.syncState();
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view1);
-        navigationView.setNavigationItemSelectedListener(this);
+        super.onCreateNavigation();
 
 
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
@@ -67,66 +57,6 @@ public class CoursePage extends AppCompatActivity implements View.OnClickListene
 
             }
         });
-    }
-
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout1);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.navi, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-
-    @SuppressWarnings("StatementWithEmptyBody")
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        // Handle navigation view item clicks here.
-        int id = item.getItemId();
-
-        if (id == R.id.nav_homepage) {
-            startActivity(new Intent(this, Homepage.class));
-
-        } else if (id == R.id.nav_course) {
-            startActivity(new Intent(this, CoursePage.class));
-
-        } else if (id == R.id.nav_4yearplan) {
-
-        } else if (id == R.id.nav_gpacalculator) {
-
-        } else if (id == R.id.nav_setting) {
-
-        } else if (id == R.id.nav_logout) {
-            startActivity(new Intent(this, Login.class));
-
-        }
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout1);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
     }
 
     @Override
