+++ MyApplication/app/src/main/res/layout/content_navi.xml
@@ -10,7 +10,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context=".Homepage"
-    tools:showIn="@layout/app_bar_navi">
+    tools:showIn="@layout/activity_navi">
 
     <TextView
         android:layout_width="wrap_content"
